+++ src/charts/index.ts
@@ -1,0 +1,19 @@
+import BarChart from './bar_chart';
+import GanttChart, { 
+    GanttChartColumnGroup as GCCG, 
+    GanttChartColumnTitle as GCCT, 
+    GanttChartContextMenuItem as GCCMI, 
+    GanttChartDataType as GCDT, 
+    GanttChartItemType as GCIT 
+} from './gantt_chart';
+
+export {
+    BarChart,
+    GanttChart
+};
+
+export type GanttChartColumnGroup = GCCG;
+export type GanttChartColumnTitle = GCCT;
+export type GanttChartContextMenuItem = GCCMI;
+export type GanttChartDataType = GCDT;
+export type GanttChartItemType = GCIT;